+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Fixes and improvements
 
+* Fix count of Arabic characters in the map of detected alphabets
+
 ## [v1.8.2](https://github.com/OpenNMT/Tokenizer/releases/tag/v1.8.2) (2018-09-10)
 
 ### Fixes and improvements